Address #200: Weird crash on dynamic cast of cache after 17000 simulation steps.
The issue had to do with free block drop when a robot aborts its task while
carrying a block. If the robot drops its block on an empty cell or one that
contains a cache, all is fine. But if it drops it on one that also contains a
block, then the cell goes into the HAS_CACHE state while still pointing to a
block.

This will be addressed in depth2, so for now I fudge it and just distribute the
block via arena_map.