Address #189: Cache penalty should affect foragers too
Implementing this feature also uncovered a rather nasty bug: If a robot aborts
its task while serving the cache penalty, it is not removed from the penalty
list, so the next time it executes a task that requires serving the penalty, it
will automatically be counted as "served", and possibly be handed a block by the
simulation that the controller is not asking for.

This was fixed, AND asserts added to safeguard against this condition in the
future.