bugfix(#429): PAM cache discrepancy on pickup
Caused by a single line typo in the line of sight calculation for the boundaries
of it, which were being used to compute if a cache overlapped the LOS, then it
should be added to it, even if the cell that actually contained the cache was
outside its extent.

This was not triggered by my verification of depth1, because I was using a 10x5
arena, in which it is much more difficult to have the LOS overlap with the cache
but not contain the cell that the actual cache resided in. It did happen a few
times, but I always dismissed it as spurious because it was so rare.

It can (more) easily be triggered if the cache is MUCH larger than the LOS, say
in a 20x10 single source scenario.

+semver: bugfix