bugfix(#432): Assert on cached block pickup
Very similar cause to #429. What can happen is that a robot sees a cache in its
LOS (and adds it to its PAM) through the "overlap" hack I put in place a while
ago. This is OK, EXCEPT when another robot makes that cache disappear, because
the current robot has no way of knowing/being informed thusly.

Therefore, it can happen that when the robot goes to pick up from/drop in a
cache shortly thereafter this occurs, that the # blocks in its PAM for the cell
containing the cache it has a quasi-dangling reference to disagrees with the #
of blocks in the real cache it is picking up from/dropping into, which has the
CORRECT # blocks in it because it is a real cache that lives in the arena.

Also implemented #379 for caches.

+semver: bugfix