bugfix(#708): COSM#151 flowdown (RN static cache segfaulting)
The root of the problem was that if a static cache is depleted and a block is
dropped onto the space it will occupy (by a task abort or by block
distribution), then when the static cache creator next runs, it will result the
state of the cell containing the block within its extent to CACHE_EXTENT, which
then results in it being removed from the RN block cluster and a segfault when
we go to remove it later. This was made worse by the unnecessary re-calculation
of block cluster membership after every block op in the loop functions (might
have been necessary in the past, but is not now).

The solution was to change the static cache creation algorithm to be more like
the dynamic cache creation algorithm, and contain a more definitive "absorb"
phase.