bugfix(#712): CRW robots stuck waiting for pickup
Caused by no block vanished event being sent, which happened ONLY if a robot was
already serving a penalty when a block disappeared. If the robot happens to
wander onto a block, signals that it is ready for a pickup, and then that same
timestep drift off the block (i.e., they just clipped the corner with one of
their ground sensors), then they will never start serving a penalty, and
therefore never get a block vanished event.

This seems to fix it for the time being. It *should* be ok for other controllers
too, but I'm not 100% sure.